513123 Scaledup producer exporter refactor (#61)
* 513123 Scaledup producer exporter refactor

* Update CalcResultsExporterTests.cs
