+++ src/EPR.Calculator.Service.Function.UnitTests/Exporter/CalcResultsExporterTests.cs
@@ -6,6 +6,7 @@
     using AutoFixture;
     using EPR.Calculator.API.Exporter;
     using EPR.Calculator.Service.Function.Exporter;
+    using EPR.Calculator.Service.Function.Exporter.ScaledupProducers;
     using EPR.Calculator.Service.Function.Models;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
@@ -16,14 +17,17 @@
     {
         private Fixture Fixture { get; } = new Fixture();
 
-        private Mock<ICalcResultDetailExporter> mockResultDetailexporter = new();
+        private Mock<ICalcResultDetailExporter> mockResultDetailExporter = new();
         private Mock<IOnePlusFourApportionmentExporter> mockOnePlusFourExporter = new();
+        private Mock<ICalcResultScaledupProducersExporter> mockScaledupProducersExporter = new();
 
         [TestMethod]
         public void Export_ShouldReturnCsvContent_WhenAllDataIsPresent()
         {
             // Arrange
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
             var calcResult = CreateCalcResult();
 
             // Act
@@ -37,7 +41,9 @@
         public void Export_DataFormatting_IsCorrect()
         {
             // Arrange
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
             var calcResult = CreateCalcResult();
 
             // Act
@@ -57,7 +63,9 @@
         {
             // Arrange
             CalcResult? results = null;
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
 
             // Act & Assert
             var ex = Assert.ThrowsException<ArgumentNullException>(() => exporter.Export(results!));
@@ -69,7 +77,9 @@
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
 
             // Act
             var result = exporter.Export(results);
@@ -83,7 +93,9 @@
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
             // Act
             var result = exporter.Export(results);
 
@@ -96,7 +108,9 @@
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
             // Act
             var result = exporter.Export(results);
 
@@ -109,7 +123,9 @@
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
             // Act
             var result = exporter.Export(results);
 
@@ -122,7 +138,9 @@
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
 
             // Act
             var result = exporter.Export(results);
@@ -132,50 +150,13 @@
         }
 
         [TestMethod]
-        public void Export_ShouldIncludeScaledupProducers_WhenNotNull()
-        {
-            // Arrange
-            var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
-
-            // Act
-            var result = exporter.Export(results);
-
-            // Assert
-            Assert.IsTrue(result.Contains("Scaled-up Producers"));
-        }
-
-        [TestMethod]
-        [DataRow(true)]
-        [DataRow(false)]
-        public void Export_ScaledUpProducer_ShouldIncludeHeadersAndDisplayNone_WhenNoScaledUpProducer(
-            bool setScaledUpProducersToNull)
-        {
-            // Arrange
-            var results = CreateCalcResult();
-            results.CalcResultScaledupProducers.ScaledupProducers = setScaledUpProducersToNull
-                ? null!
-                : new List<CalcResultScaledupProducer>();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
-
-            // Act
-            var result = exporter.Export(results);
-
-            // Assert
-            Assert.IsTrue(result.Contains("Scaled-up Producers"));
-            Assert.IsTrue(result.Contains("Each submission for the year"));
-            Assert.IsTrue(result.Contains("Aluminium Breakdown"));
-            Assert.IsTrue(result.Contains("Producer ID"));
-            Assert.IsTrue(result.Contains("Subsidiary ID"));
-            Assert.IsTrue(result.Contains("None"));
-        }
-
-        [TestMethod]
         public void Export_ShouldIncludeSummaryData_WhenNotNull()
         {
             // Arrange
             var results = CreateCalcResult();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
 
             // Act
             var result = exporter.Export(results);
@@ -194,7 +175,9 @@
                 CalcResultLateReportingTonnageData = null!,
                 CalcResultParameterOtherCost = null!,
             };
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
+            var exporter = new CalcResultsExporter(mockResultDetailExporter.Object,
+                mockOnePlusFourExporter.Object,
+                mockScaledupProducersExporter.Object);
 
             // Act
             if (results != null)
@@ -217,22 +200,6 @@
                     "CSV content should not contain LaDisposalCostData.");
                 Assert.IsFalse(csvContent.Contains("SummaryData"), "CSV content should not contain SummaryData.");
             }
-        }
-
-        [TestMethod]
-        public void Export_ShouldIncludeGlassColumns_WhenGlassMaterialPresent()
-        {
-            // Arrange
-            var results = CreateCalcResultWithGlass();
-            var exporter = new CalcResultsExporter(mockResultDetailexporter.Object, mockOnePlusFourExporter.Object);
-
-            // Act
-            var result = exporter.Export(results);
-
-            // Assert
-            Assert.IsTrue(result.Contains("Glass"));
-            Assert.IsTrue(result.Contains("HouseholdDrinksContainersTonnageGlass"));
-            Assert.IsTrue(result.Contains("ScaledupHouseholdDrinksContainersTonnageGlass"));
         }
 
         [TestMethod]
@@ -560,46 +527,6 @@
             return tonnageByMaterial;
         }
 
-        private static CalcResult CreateCalcResultWithGlass()
-        {
-            var result = CreateCalcResult();
-            result.CalcResultScaledupProducers.ScaledupProducers = GetCalcResultScaledupProducerListWithGlass();
-            return result;
-        }
-
-        private static IEnumerable<CalcResultScaledupProducer> GetCalcResultScaledupProducerListWithGlass()
-        {
-            var scaledupProducerList = new List<CalcResultScaledupProducer>
-            {
-                new CalcResultScaledupProducer
-                {
-                    ProducerId = 101001,
-                    SubsidiaryId = string.Empty,
-                    ProducerName = "Allied Packaging",
-                    Level = "1",
-                    SubmissionPeriodCode = "2024-P2",
-                    DaysInSubmissionPeriod = 91,
-                    DaysInWholePeriod = 91,
-                    ScaleupFactor = 2,
-                    ScaledupProducerTonnageByMaterial = GetScaledupProducerTonnageByMaterialWithGlass(),
-                },
-                new CalcResultScaledupProducer
-                {
-                    ProducerId = 101001,
-                    SubsidiaryId = string.Empty,
-                    ProducerName = "Allied Packaging",
-                    Level = "1",
-                    SubmissionPeriodCode = "2024-P2",
-                    DaysInSubmissionPeriod = 91,
-                    DaysInWholePeriod = 91,
-                    ScaleupFactor = 2,
-                    ScaledupProducerTonnageByMaterial = GetScaledupProducerTonnageByMaterialWithGlass(),
-                    IsTotalRow = true,
-                },
-            };
-            return scaledupProducerList;
-        }
-
         private static List<CalcResultScaledupProducer> GetCalcResultScaledupProducerList()
         {
             var scaledupProducerList = new List<CalcResultScaledupProducer>();
@@ -633,32 +560,6 @@
             ]);
 
             return scaledupProducerList;
-        }
-
-        private static Dictionary<string, CalcResultScaledupProducerTonnage> GetScaledupProducerTonnageByMaterialWithGlass()
-        {
-            var tonnageByMaterial = new Dictionary<string, CalcResultScaledupProducerTonnage>
-            {
-                {
-                    "GL",
-                    new CalcResultScaledupProducerTonnage
-                    {
-                        ReportedHouseholdPackagingWasteTonnage = 1000,
-                        ReportedPublicBinTonnage = 100,
-                        HouseholdDrinksContainersTonnageGlass = 50,
-                        TotalReportedTonnage = 1100,
-                        ReportedSelfManagedConsumerWasteTonnage = 500,
-                        NetReportedTonnage = 1100,
-                        ScaledupReportedHouseholdPackagingWasteTonnage = 2000,
-                        ScaledupReportedPublicBinTonnage = 200,
-                        ScaledupHouseholdDrinksContainersTonnageGlass = 100,
-                        ScaledupTotalReportedTonnage = 2200,
-                        ScaledupReportedSelfManagedConsumerWasteTonnage = 1000,
-                        ScaledupNetReportedTonnage = 2200,
-                    }
-                },
-            };
-            return tonnageByMaterial;
         }
     }
 }
